(ft-test-endpoints), added send query test
